added button disabled on click

+++ main.py
@@ -14,23 +14,24 @@ eel.init("web");
 
 @eel.expose
 def windowfilepicker():
-    for file in os.listdir("web/images/scanned_receipts"):
-        os.remove("web/images/scanned_receipts/{}".format(file))
+    directory="web/images/parsed_receipts/";
     tkinter.Tk().withdraw(); # prevents an empty tkinter window from appearing
     file_paths = list(filedialog.askopenfilenames());
-    #use os.replace instead
-    file_names = [];
-    for index, path in enumerate(file_paths):
-        image_mat = cv.imread(path);
-        file_names.append(pattern.findall(path)[0])
-        cv.imwrite("web/images/scanned_receipts/{}".format(file_names[index]),image_mat);
-    eel.display_images(file_names);
+    if(file_paths):
+        for file in os.listdir(directory):
+            os.remove(directory+file)
+        for file in os.listdir("web/images/scanned_receipts"):
+            os.remove("web/images/scanned_receipts/{}".format(file))
+        #use os.replace instead
+        file_names = [];
+        for index, path in enumerate(file_paths):
+            image_mat = cv.imread(path);
+            file_names.append(pattern.findall(path)[0])
+            cv.imwrite("web/images/scanned_receipts/{}".format(file_names[index]),image_mat);
+        eel.display_images(file_names);
 
 @eel.expose
 def parse():
-    directory="web/images/parsed_receipts/";
-    for file in os.listdir(directory):
-        os.remove(directory+file)
     correspondence = parse_receipts();
     eel.display_parsed_images(correspondence);
 
@@ -38,6 +39,6 @@ def parse():
 def ocr():
     # os.remove("expenses.xlsx")
     amazon_ocr();
-    # print("done")
+    eel.reable_excel_button();    
 
 eel.start("index.html");
